ALPHA 1.02
Bug fixes (code review)
- Fix const-reassignment crash in scheduleWeek swap (now reachable after the
  swapIndexes rename)
- Allow editing dynamic events via /magic (no longer rejected as "missing date")
- Recover dynamic-overlap fixes against the correct ISO week
- Build assistant action dates via setHours to avoid Invalid Date on single-digit
  hours; preserve custom event colour on edits; merge tags instead of replacing
- Render multi-group chat headers correctly (name/groupName)

Database
- Make EventDate.countWeek a Number, Message.sessionId an ObjectId ref
- Add indexes (Message, Session), unique membership index, OAuth/Session TTLs
- Fix `required` typos and drop the no-op `timeseries` option

Assistant features
- /magic now handles multiple events per request, each confirmed individually
  (resolveMagicActions; controller returns { messages, actions }; MessageList
  renders one confirmable card per action with its own done/cancelled state)
- Analyzer proposes concrete fixes: overlaps carry a shift-time suggestion that
  the analyze endpoint resolves into confirmable edit actions
- New /organizer command suggests relevant tags per event (organizerPrompt +
  getGroqOrganize + controller.organize + POST /message/organize)

Permissions
- Add configurable group parameter `assistantCommandRole` (default admin)
- Enforce on both ends: backend checkGroupParamRole on event-write routes +
  per-group gateActions in the message controller; frontend hides/blocks
  /magic and /organizer when the user's role doesn't meet the threshold
  (/analyze stays available to all viewers)

Schedule sync after assistant edits
- ModalAssitent reports applied changes via onScheduleMutated; SchedulePage
  invalidates the affected group's cache so the view never shows stale data
- Improved refresh check (isChangeRelevant): a single relevant change refreshes
  silently; several show a ConfirmModal asking before reloading; off-screen
  group/week changes only invalidate cache without prompting

UI
- Redesign the assistant scope panel into an inline pink filter row
- Fix RangeDatePicker truncation (short date format) and popup positioning

+++ Front-end/src/UI/RangeDatePicker/RangeDatePicker.js
@@ -25,7 +25,8 @@ const RangeDatePicker = ({
     const toDateVal = useMemo(() => toDate(to), [to, toDate]);
 
     const display = useMemo(() => {
-        const opts = { year: 'numeric', month: 'long', day: 'numeric' };
+        // Short month keeps a full "from – to" range readable in tight inline layouts.
+        const opts = { year: 'numeric', month: 'short', day: 'numeric' };
         const f = fromDate ? fromDate.toLocaleDateString(undefined, opts) : '';
         const tt = toDateVal ? toDateVal.toLocaleDateString(undefined, opts) : '';
         if (f && tt) return `${f} – ${tt}`;
